soon and ignore some mistake
OK, it will be more strong when some mistakes happen. Add new function "bool isMoving()". It will tell us if the manipulate is moving. So we won't cost much extra-time for waiting manipulate. And if manipulate stop moving but do not arrive the end-point, it will rewrite end-point to manipulate without any trajectory.